Implement arrays on separate files
